+++ server/app/adminapi/logic/WorkbenchLogic.php
@@ -26,7 +26,7 @@ class WorkbenchLogic extends BaseLogic
     {
         $website = self::websiteService()->get();
         return [
-            'version' => (string) config('project.version', '1.1.5'),
+            'version' => (string) config('project.version', '2.0.0'),
             'website' => $website['official_url'],
             'name'    => $website['name'],
             'based'   => (string) config(